# Speed up `test_mxfp_matmul` by reducing block sizes: 128->64