I ran the test twice, for 128 and 64 blocks, to see the impact of compilation and +- pure runtime, and got the following data:

128 blocks:
1. `192 passed, 96 xfailed in 1464.76s (0:24:24)`
2. `192 passed, 96 xfailed in 7.16s`

64 blocks:
1. `192 passed, 96 xfailed in 235.29s (0:03:55)`
3. `192 passed, 96 xfailed in 5.99s`

It can be reverted in the future when compilation works better.